// src/types/CalendarEventType.ts
export type CalendarEventType = {
    // id: number;
    title: string;
    // description? : string;
    start: Date ;
    end: Date;
    // background_color?: string;
    // text_color?: string;
    allDay: boolean;
    // extendedProps: {
    //     student: {
    //         id: number;
    //         first_name: string;
    //         last_name: string;
    //     };
    // };
}